+++ test/e2e/specs/site-editor/user-global-styles-revisions.spec.js
@@ -80,7 +80,7 @@ test.describe( 'Style Revisions', () => {
 			.getByRole( 'button', { name: 'Background', exact: true } )
 			.click();
 		await page
-			.getByRole( 'option', { name: 'Color: Luminous vivid amber' } )
+			.getByRole( 'option', { name: 'Luminous vivid amber' } )
 			.click( { force: true } );
 
 		await page.getByRole( 'button', { name: 'Revisions' } ).click();